refactor(extension-runtime): slack preference-target codec moves to the adapter crate (P6 S6.4)
PreferenceTargetCodec (the vendor half of the triggered-delivery driver)
relocates from ironclaw_product_workflow to ironclaw_product_adapters —
the contract crate channel extension crates may depend on — with a
compat re-export in place. The slack binding-ref grammar (encode,
decode, personal-DM predicate) and SlackPreferenceTargetCodec move into
ironclaw_slack_v2_adapter::preference_targets with their pure tests;
the retiring composition lane keeps thin error-mapping wrappers until
it deletes.
